IGNITE-23683 / move LeaseInfo and its serializer to make them globally available
